- integration tests for our apis - component tests for services - h2 in memory db for test profile - fixed initial schema (Int -> BigInt for PKs) - added rest assured and h2 libs (for tests only) - updated gradle task config for test logs

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -1,17 +1,17 @@
 CREATE TABLE Accounts (
-    Account_ID INT AUTO_INCREMENT PRIMARY KEY,
+    Account_ID BIGINT AUTO_INCREMENT PRIMARY KEY,
     Document_Number VARCHAR(45)
 );
 
 CREATE TABLE OperationsTypes (
-    OperationType_ID INT AUTO_INCREMENT PRIMARY KEY,
+    OperationType_ID BIGINT AUTO_INCREMENT PRIMARY KEY,
     Description VARCHAR(100)
 );
 
 CREATE TABLE Transactions (
-    Transaction_ID INT AUTO_INCREMENT PRIMARY KEY,
-    Account_ID INT,
-    OperationType_ID INT,
+    Transaction_ID BIGINT AUTO_INCREMENT PRIMARY KEY,
+    Account_ID BIGINT,
+    OperationType_ID BIGINT,
     Amount DECIMAL(19, 4) NOT NULL,
     EventDate DATETIME(6) NOT NULL,
     FOREIGN KEY (Account_ID) REFERENCES Accounts(Account_ID),
